requirements.txt
Django==1.6
South==0.8.3

# Update pip when possible
pip

ipdb
python-dateutil==2.2
django-qurl==0.1.1